pyproject.toml
[project]
name = "propertyMaster"
version = "0.1.0"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "pandas>=2.3.1",
    "requests>=2.32.4",
    "simpledbf>=0.2.6",
    "sqlalchemy>=2.0.42",
    "python-dotenv==1.1.1",
]
